chore: refactor task controller
