--- src/app_web/controllers/task/index.ts
@@ -1,1 +1,0 @@
-export * from "./task";
